+++ formidable/__init__.py
@@ -3,5 +3,5 @@ from __future__ import absolute_import
 from .json_migrations import latest_version
 
 default_app_config = 'formidable.app.FormidableConfig'
-version = '1.5.1'
+version = '1.6.0.dev0'
 json_version = latest_version