+++ formidable/serializers/forms.py
@@ -29,7 +29,11 @@ class ConditionTestSerializer(serializers.Serializer):
 
 
 class ConditionSerializer(serializers.Serializer):
-    name = serializers.CharField(required=False, allow_null=True)
+    name = serializers.CharField(
+        required=False,
+        allow_null=True,
+        allow_blank=True,
+    )
     fields_ids = serializers.ListField(
         child=serializers.CharField(max_length=256),
         allow_empty=False,